Extract PONativeAlternativePaymentMethodParameterValues to separate file as it's shared
